Added fix to networking issues in Windows Docker
The following applies to line 3
The -f will force close any open indra instances
and rm will remove it like normal

The following applies to lines 9-13.
This is for forwarding the VM from its internal network(which can have it's own localhost) to the host machine's local loopback.
This cannot be generalized to be a docker command (nor a docker-machine) command
However it can be put into a VBoxManage command.
Specifically --> VBoxManage modifyvm " <DOCKER MACHINE IMAGE NAME> " --natpf1 "django,tcp,0.0.0.0,8000,,8000"
But this would generate an error on machines that don't have VBoxManage as an alias on their powershell
Also would generate an error where VM's arent used

The following applies to lines 21 and 22
The absolute path used by docker has to be formatted to work properly. $pwd.path (the string version of $pwd) has a captial volume identifier followed by a colon and uses backslashes like C:\...
Docker prefers /c/...
so I glued together a /, the lowercase version of the first letter
of $pwd, and the part of $pwd past the colon